+++ tests/config/test_get_coltrane_settings.py
@@ -1,6 +1,12 @@
 import pytest
 
-from coltrane.config.settings import DEFAULT_COLTRANE_SETTINGS, get_coltrane_settings
+from coltrane.config.settings import (
+    DEFAULT_COLTRANE_SETTINGS,
+    get_coltrane_settings,
+    get_description,
+    get_site,
+    get_title,
+)
 
 
 def test_get_coltrane_settings_none(settings):
@@ -18,3 +24,18 @@ def test_get_coltrane_settings_invalid_type(settings):
 
     with pytest.raises(AssertionError):
         get_coltrane_settings()
+
+
+def test_get_site(settings):
+    setattr(settings, "COLTRANE", {"SITE": "http://test-site.com"})
+    assert get_site() == "http://test-site.com"
+
+
+def test_get_title(settings):
+    setattr(settings, "COLTRANE", {"TITLE": "test-title"})
+    assert get_title() == "test-title"
+
+
+def test_get_description(settings):
+    setattr(settings, "COLTRANE", {"DESCRIPTION": "test description"})
+    assert get_description() == "test description"