+++ tests/urls/test_urlpatterns.py
@@ -3,12 +3,13 @@ import importlib
 from coltrane import urls
 
 
-def test_urlpatterns_has_2():
-    assert len(urls.urlpatterns) == 2
+def test_urlpatterns():
+    importlib.reload(urls)
+    assert len(urls.urlpatterns) == 3
 
 
-def test_debug_urlpatterns_has_3(settings):
+def test_debug_urlpatterns(settings):
     settings.DEBUG = True
     importlib.reload(urls)
 
-    assert len(urls.urlpatterns) == 3
+    assert len(urls.urlpatterns) == 4